Added SimulatorCommunicationsChannel and support in the DeviceControllerContextBuilder class for creating tests that use the hardware simulator. This required changes to the Reactive Communications library to allow the user to register custom channels.

+++ TA.DigitalDomeworks.SharedTypes/ControllerStatus.cs
@@ -211,7 +211,7 @@ namespace TA.DigitalDomeworks.SharedTypes
                 AtHome ? 0 : 1,
                 HomeCounterClockwise,
                 HomeClockwise,
-                UserPins,
+                (byte)UserPins,
                 WeatherAge,
                 WindDirection,
                 WindSpeed,
